requirements.txt
alembic==1.16.4
astroid==3.3.11
blinker==1.9.0
click==8.2.1
dill==0.4.0
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.8
psycopg2==2.9.10
pylint==3.3.7
python-dotenv==1.1.1
SQLAlchemy==2.0.41
tomlkit==0.13.3
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1